requirements.txt
python-telegram-bot==21.8
python-dotenv==1.0.0
openai==2.8.1
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2
qdrant-client==1.16.0
pdfplumber==0.10.4
